# pkg/client/openapi.yaml
components:
  schemas:
...
    AddCommentRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/AddCommentRequest.json
          format: uri
          readOnly: true
          type: string
        comment:
          type: string
        commenter:
          type: string
        job_id:
          format: int64
          type: integer
        sha:
          type: string
      required:
        - commenter
        - comment
      type: object
...
    CancelJobOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/CancelJobOutputBody.json
          format: uri
          readOnly: true
          type: string
        success:
          type: boolean
      required:
        - success
      type: object
    CancelJobRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/CancelJobRequest.json
          format: uri
          readOnly: true
          type: string
        job_id:
          format: int64
          type: integer
      required:
        - job_id
      type: object
    CloseReviewOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/CloseReviewOutputBody.json
          format: uri
          readOnly: true
          type: string
        success:
          type: boolean
      required:
        - success
      type: object
    CloseReviewRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/CloseReviewRequest.json
          format: uri
          readOnly: true
          type: string
        closed:
          type: boolean
        job_id:
          format: int64
          type: integer
      required:
        - job_id
        - closed
      type: object
...
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    ListBranchesOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/ListBranchesOutputBody.json
          format: uri
          readOnly: true
          type: string
        branches:
          items:
            $ref: "#/components/schemas/BranchWithCount"
          type:
            - array
            - "null"
        nulls_remaining:
          format: int64
          type: integer
        total_count:
          format: int64
          type: integer
      required:
        - branches
        - total_count
        - nulls_remaining
      type: object
    ListCommentsOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/ListCommentsOutputBody.json
          format: uri
          readOnly: true
          type: string
        responses:
          items:
            $ref: "#/components/schemas/Response"
          type:
            - array
            - "null"
      required:
        - responses
      type: object
    ListJobsOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/ListJobsOutputBody.json
          format: uri
          readOnly: true
          type: string
        has_more:
          type: boolean
        jobs:
          items:
            $ref: "#/components/schemas/ReviewJob"
          type:
            - array
            - "null"
        stats:
          $ref: "#/components/schemas/JobStats"
      required:
        - jobs
        - has_more
      type: object
    ListReposOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/ListReposOutputBody.json
          format: uri
          readOnly: true
          type: string
        repos:
          items:
            $ref: "#/components/schemas/RepoWithCount"
          type:
            - array
            - "null"
        total_count:
          format: int64
          type: integer
      required:
        - repos
        - total_count
      type: object
...
    RerunJobOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/RerunJobOutputBody.json
          format: uri
          readOnly: true
          type: string
        success:
          type: boolean
      required:
        - success
      type: object
    RerunJobRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/RerunJobRequest.json
          format: uri
          readOnly: true
          type: string
        job_id:
          format: int64
          type: integer
      required:
        - job_id
      type: object
...
    Response:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/Response.json
          format: uri
          readOnly: true
          type: string
        commit_id:
          format: int64
          type: integer
        created_at:
          format: date-time
          type: string
        id:
          format: int64
          type: integer
        job_id:
          format: int64
          type: integer
        responder:
          type: string
        response:
          type: string
        source_machine_id:
          type: string
        synced_at:
          format: date-time
          type: string
        uuid:
          type: string
      required:
        - id
        - responder
        - response
        - created_at
      type: object
    Review:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/Review.json
          format: uri
          readOnly: true
          type: string
        agent:
          type: string
        closed:
          type: boolean
        created_at:
          format: date-time
          type: string
        id:
          format: int64
          type: integer
        job:
          $ref: "#/components/schemas/ReviewJob"
        job_id:
          format: int64
          type: integer
        output:
          type: string
        prompt:
          type: string
        synced_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
        updated_by_machine_id:
          type: string
        uuid:
          type: string
        verdict_bool:
          format: int64
          type: integer
      required:
        - id
        - job_id
        - agent
        - prompt
        - output
        - created_at
        - closed
      type: object
...
    WebBootstrapInputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/WebBootstrapInputBody.json
          format: uri
          readOnly: true
          type: string
      type: object
    WebLoginRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/WebLoginRequest.json
          format: uri
          readOnly: true
          type: string
        token:
          minLength: 1
          type: string
      required:
        - token
      type: object
    WebSessionCredentials:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/WebSessionCredentials.json
          format: uri
          readOnly: true
          type: string
        csrf:
          type: string
        expires_at:
          format: date-time
          type: string
        session:
          type: string
      required:
        - session
        - csrf
        - expires_at
      type: object
    WebSessionStatus:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/WebSessionStatus.json
          format: uri
          readOnly: true
          type: string
        authenticated:
          type: boolean
        authentication:
          enum:
            - local
            - token
          type: string
        expires_at:
          format: date-time
          type: string
      required:
        - authentication
        - authenticated
      type: object
info:
  title: roborev
  version: dev
openapi: 3.1.0
paths:
...
  /api/branches:
    get:
      operationId: list-branches
      parameters:
        - description: Filter to branches in these repo paths
          explode: true
          in: query
          name: repo
          schema:
            description: Filter to branches in these repo paths
            items:
              type: string
            type:
              - array
              - "null"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListBranchesOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List branches with job counts
      tags:
        - repos
  /api/comment:
    post:
      operationId: add-comment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Add a comment to a job or commit
      tags:
        - comments
  /api/comments:
    get:
      operationId: list-comments
      parameters:
        - description: List comments by job ID
          explode: false
          in: query
          name: job_id
          schema:
            default: -1
            description: List comments by job ID
            format: int64
            type: integer
        - description: List comments by commit ID
          explode: false
          in: query
          name: commit_id
          schema:
            default: -1
            description: List comments by commit ID
            format: int64
            type: integer
        - description: List comments by commit SHA
          explode: false
          in: query
          name: sha
          schema:
            description: List comments by commit SHA
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCommentsOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List comments for a job or commit
      tags:
        - comments
...
  /api/job/cancel:
    post:
      operationId: cancel-job
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelJobRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelJobOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Cancel a queued or running job
      tags:
        - jobs
...
  /api/job/log:
    get:
      operationId: get-job-log
      parameters:
        - description: Job ID
          explode: false
          in: query
          name: job_id
          schema:
            description: Job ID
            type: string
        - description: Byte offset into the log file
          explode: false
          in: query
          name: offset
          schema:
            description: Byte offset into the log file
            type: string
      responses:
        "200":
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get raw job log bytes
      tags:
        - jobs
  /api/job/output:
    get:
      operationId: get-job-output
      parameters:
        - description: Job ID
          explode: false
          in: query
          name: job_id
          schema:
            description: Job ID
            type: string
        - description: Stream output as NDJSON when set to 1
          explode: false
          in: query
          name: stream
          schema:
            description: Stream output as NDJSON when set to 1
            type: string
      responses:
        "200":
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get or stream in-memory job output
      tags:
        - jobs
...
  /api/job/rerun:
    post:
      operationId: rerun-job
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RerunJobRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RerunJobOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Re-enqueue a completed or failed job
      tags:
        - jobs
...
  /api/jobs:
    get:
      operationId: list-jobs
      parameters:
        - description: Return a single job by ID
          explode: false
          in: query
          name: id
          schema:
            default: -1
            description: Return a single job by ID
            format: int64
            type: integer
        - description: Filter by job status
          explode: false
          in: query
          name: status
          schema:
            description: Filter by job status
            type: string
        - description: Filter by repo root path (repeatable)
          explode: true
          in: query
          name: repo
          schema:
            description: Filter by repo root path (repeatable)
            items:
              type: string
            type:
              - array
              - "null"
        - description: Filter by git ref
          explode: false
          in: query
          name: git_ref
          schema:
            description: Filter by git ref
            type: string
        - description: Filter by branch name
          explode: false
          in: query
          name: branch
          schema:
            description: Filter by branch name
            type: string
        - description: Include jobs with no branch when filtering by branch
          explode: false
          in: query
          name: branch_include_empty
          schema:
            description: Include jobs with no branch when filtering by branch
            enum:
              - "true"
              - "false"
              - ""
            type: string
        - description: Filter by review closed state
          explode: false
          in: query
          name: closed
          schema:
            description: Filter by review closed state
            enum:
              - "true"
              - "false"
              - ""
            type: string
        - description: Filter by job type
          explode: false
          in: query
          name: job_type
          schema:
            description: Filter by job type
            type: string
        - description: Exclude jobs of this type
          explode: false
          in: query
          name: exclude_job_type
          schema:
            description: Exclude jobs of this type
            type: string
        - description: Hide auto-design-router rows (job_type=classify and status=skipped)
          explode: false
          in: query
          name: hide_classify_jobs
          schema:
            description: Hide auto-design-router rows (job_type=classify and status=skipped)
            enum:
              - "true"
              - "false"
              - ""
            type: string
        - description: Return all jobs (members + synthesis) of one panel run
          explode: false
          in: query
          name: panel_run
          schema:
            description: Return all jobs (members + synthesis) of one panel run
            type: string
        - description: Omit prompt and diff content from returned jobs (metadata-only listing; queued/running jobs keep their prompt)
          explode: false
          in: query
          name: omit_prompt
          schema:
            description: Omit prompt and diff content from returned jobs (metadata-only listing; queued/running jobs keep their prompt)
            enum:
              - "true"
              - "false"
              - ""
            type: string
        - description: Filter repos by path prefix
          explode: false
          in: query
          name: repo_prefix
          schema:
            description: Filter repos by path prefix
            type: string
        - description: Max results (default 50, 0=unlimited, max 10000)
          explode: false
          in: query
          name: limit
          schema:
            default: -999999
            description: Max results (default 50, 0=unlimited, max 10000)
            format: int64
            type: integer
        - description: Skip N results (requires limit>0)
          explode: false
          in: query
          name: offset
          schema:
            default: -1
            description: Skip N results (requires limit>0)
            format: int64
            type: integer
        - description: "Cursor: return jobs with ID < this value"
          explode: false
          in: query
          name: before
          schema:
            default: -1
            description: "Cursor: return jobs with ID < this value"
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListJobsOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List review jobs
      tags:
        - jobs
...
  /api/repos:
    get:
      operationId: list-repos
      parameters:
        - description: Filter to repos with jobs on this branch
          explode: false
          in: query
          name: branch
          schema:
            description: Filter to repos with jobs on this branch
            type: string
        - description: Filter repos by path prefix
          explode: false
          in: query
          name: prefix
          schema:
            description: Filter repos by path prefix
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReposOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List repos with job counts
      tags:
        - repos
...
  /api/review:
    get:
      operationId: get-review
      parameters:
        - description: Look up review by job ID
          explode: false
          in: query
          name: job_id
          schema:
            default: -1
            description: Look up review by job ID
            format: int64
            type: integer
        - description: Look up review by commit SHA
          explode: false
          in: query
          name: sha
          schema:
            description: Look up review by commit SHA
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get a review by job ID or SHA
      tags:
        - reviews
  /api/review/close:
    post:
      operationId: close-review
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseReviewRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloseReviewOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Close or reopen a review
      tags:
        - reviews
...
  /api/status:
    get:
      operationId: get-status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DaemonStatus"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get daemon status
      tags:
        - daemon
  /api/stream/events:
    get:
      operationId: stream-events
      parameters:
        - description: Filter events by repo root path
          explode: false
          in: query
          name: repo
          schema:
            description: Filter events by repo root path
            type: string
      responses:
        "200":
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Stream daemon events
      tags:
        - daemon
...
  /api/ui/session:
    delete:
      operationId: logout-web-session
      responses:
        "204":
          description: No Content
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Invalidate a browser session
      tags:
        - web-session
    get:
      operationId: get-web-session-status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebSessionStatus"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get browser authentication status
      tags:
        - web-session
  /api/ui/session/bootstrap:
    post:
      operationId: bootstrap-web-session
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebBootstrapInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebSessionCredentials"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Mint tab credentials from an ambient browser session
      tags:
        - web-session
  /api/ui/session/login:
    post:
      operationId: login-web-session
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebLoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebSessionCredentials"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Exchange a daemon token for a browser session
      tags:
        - web-session